+++ tst/unit/test_swarm.cpp
@@ -3,7 +3,7 @@
 // Copyright(C) 2020 The Parthenon collaboration
 // Licensed under the 3-clause BSD License, see LICENSE file for details
 //========================================================================================
-// (C) (or copyright) 2020-2022. Triad National Security, LLC. All rights reserved.
+// (C) (or copyright) 2020-2023. Triad National Security, LLC. All rights reserved.
 //
 // This program was produced under U.S. Government contract 89233218CNA000001
 // for Los Alamos National Laboratory (LANL), which is operated by Triad
@@ -59,11 +59,6 @@ class ParticleBoundIX1User : public ParticleBound {
   }
 };
 
-std::unique_ptr<ParticleBound, DeviceDeleter<parthenon::DevMemSpace>>
-SetSwarmIX1UserBC() {
-  return DeviceAllocate<ParticleBoundIX1User>();
-}
-
 TEST_CASE("Swarm memory management", "[Swarm]") {
   std::stringstream is;
   is << "<parthenon/mesh>" << endl;
@@ -76,17 +71,20 @@ TEST_CASE("Swarm memory management", "[Swarm]") {
   is << "nx1 = 4" << endl;
   is << "nx2 = 4" << endl;
   is << "nx3 = 4" << endl;
+  is << "swarm_ix1_bc = user" << endl;
+  is << "swarm_ox1_bc = outflow" << endl;
+  is << "swarm_ix2_bc = outflow" << endl;
+  is << "swarm_ox2_bc = outflow" << endl;
+  is << "swarm_ix3_bc = outflow" << endl;
+  is << "swarm_ox3_bc = outflow" << endl;
   auto pin = std::make_shared<ParameterInput>();
   pin->LoadFromStream(is);
   auto app_in = std::make_shared<ApplicationInput>();
+  app_in->RegisterSwarmBoundaryCondition<ParticleBoundIX1User>(
+      parthenon::BoundaryFace::inner_x1);
   Packages_t packages;
   auto meshblock = std::make_shared<MeshBlock>(1, 1);
   auto mesh = std::make_shared<Mesh>(pin.get(), app_in.get(), packages, 1);
-  mesh->mesh_bcs[0] = BoundaryFlag::user;
-  mesh->SwarmBndryFnctn[0] = SetSwarmIX1UserBC;
-  for (int i = 1; i < 6; i++) {
-    mesh->mesh_bcs[i] = BoundaryFlag::outflow;
-  }
   meshblock->pmy_mesh = mesh.get();
   Metadata m;
   auto swarm = std::make_shared<Swarm>("test swarm", m, NUMINIT);